+ Se agregan validaciones al metodo de captura de eventos de mouse. + Se agrega funcionalidad a la grilla y se prueba minimamente en el modo DDA

+++ trunk/tp1/Geometria/Circunferencia.cpp
@@ -71,7 +71,7 @@ void Circunferencia::rellenar() {
     glColor3f(this->relleno->getRojo(), this->relleno->getVerde(), this->relleno->getAzul());
 }
 
-bool Circunferencia::contiene(float x, float y){
+bool Circunferencia::contiene(int x, int y){
     //TODO
     return false;
 }
