+++ trunk/tp1/Geometria/Circunferencia.h
@@ -35,7 +35,7 @@ public:
     /**
      * Devuelve true si la figura contiene al punto (x,y)
      */
-    bool contiene(float x, float y);
+    bool contiene(int x, int y);
 
 protected:
 private:
