+++ trunk/tp1/Geometria/Rectangulo.h
@@ -43,7 +43,7 @@ class Rectangulo : public FiguraGeometrica
         /**
          * Devuelve true si la figura contiene al punto (x,y)
          */
-        bool contiene(float x, float y);
+        bool contiene(int x, int y);
 
         /** Sobrecarga del operador = **/
         bool operator = (Rectangulo* rectangulo);
